Add auth support to rtsp:// netcam
The ffmpeg code expects the auth info to be part of the url that is
passed to it. This constructs the url using either the netcam_userpass
configuration option, or if that is missing from the auth data
parsed from the netcam_url option.

This also adds a call to avformat_network_init() in rtsp_connect to
quiet a deprecation warning.